refactor: fix modal types and helper file name
