+++ src/helpers/convertRulesRegExp.ts
