+++ src/helpers/modalFunctions.ts
@@ -1,6 +1,8 @@
-import { ModalAction, ModalConfirmAction } from '../types/modal';
+import { ModalActionKey, ModalConfirmAction } from '../types/modal';
 import { modalConfirmAction } from '../constants/modal';
 
-export default function isConfirmAction(action: ModalAction): action is keyof ModalConfirmAction {
+export default function isConfirmAction(
+  action: ModalActionKey,
+): action is keyof ModalConfirmAction {
   return (action as keyof typeof modalConfirmAction) in modalConfirmAction;
 }
